+++ katana/server.py
@@ -89,7 +89,7 @@ class ComponentServer(object):
 
         self.__args = args
         self.__socket = None
-        self.__schema_registry = get_schema_registry()
+        self.__registry = get_schema_registry()
         self._pool = ThreadPool(cpu_count() * 5)
 
         self.callbacks = callbacks
@@ -207,7 +207,7 @@ class ComponentServer(object):
 
         LOG.debug('Updating schemas for Services ...')
         try:
-            self.__schema_registry.update_registry(unpack(stream))
+            self.__registry.update_registry(unpack(stream))
         except:
             LOG.exception('Failed to update schemas')