+++ katana/api/base.py
@@ -224,4 +224,6 @@ class Api(object):
 
         """
 
-        return False
+        raise ApiError(
+            'SDK does not support async call to end action: Api.done()'
+            )