+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- A configured file server must be available to make runtime calls
+  that send file parameters.
 
 ## [1.3.0] - 2017-11-01
 ### Added